intent data

+++ app/src/main/java/za/ac/tuthospitalmanagementsystem/RecordActivity.kt
@@ -3,10 +3,7 @@ package za.ac.tuthospitalmanagementsystem
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
-import android.widget.TextView
-import android.widget.Toast
+import android.widget.*
 import androidx.appcompat.widget.Toolbar
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
@@ -15,30 +12,54 @@ import java.lang.StringBuilder
 class RecordActivity : AppCompatActivity() {
     private lateinit var database : DatabaseReference
     private lateinit var recordNo : String
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_record)
-        var toolbar = findViewById<Toolbar>(R.id.toolbar)
+        val toolbar = findViewById<Toolbar>(R.id.toolbar)
+
 
         setSupportActionBar(toolbar)
         supportActionBar!!.title = "Record"
         val buttonEdit = findViewById<Button>(R.id.buttonEdit)
         val buttonDelete = findViewById<Button>(R.id.buttonDelete)
+
+        val name: String = intent.getStringExtra("name").toString()
+        val surname: String= intent.getStringExtra("surname").toString()
+        val number: String= intent.getStringExtra("number").toString()
+        val username: String= intent.getStringExtra("username").toString()
+
         viewAllRecords()
         buttonEdit.setOnClickListener {
-            goToEdit()
+            goToEdit(username,name,surname,number)
         }
         buttonDelete.setOnClickListener {
-            deleteRecord()
+            deleteRecord(username,name,surname,number)
         }
         viewAllRecords()
     }
 
-    private fun deleteRecord() {
+
+    private fun deleteRecord(username: String, name: String, surname: String, number: String) {
         recordNo = findViewById<EditText>(R.id.editTextRecordNo).text.toString()
         database = FirebaseDatabase.getInstance().getReference("PatientRecord").child(recordNo)
-        database.removeValue().addOnSuccessListener {
-            Toast.makeText(this,"Record removed",Toast.LENGTH_LONG).show()
+        database.get().addOnSuccessListener {
+            if(it.exists()){
+                database.removeValue().addOnSuccessListener {
+                    Toast.makeText(this,"Record removed",Toast.LENGTH_LONG).show()
+                }
+                val intent = Intent(this,DoctorActivity::class.java)
+                intent.putExtra("recordNo",recordNo)
+                intent.putExtra("username",username)
+                intent.putExtra("name",name)
+                intent.putExtra("surname",surname)
+                intent.putExtra("number",number)
+                startActivity(intent)
+            }else{
+                Toast.makeText(this,"Patient does not exist",Toast.LENGTH_LONG).show()
+            }
+        }.addOnFailureListener {
+            Toast.makeText(this,"failed", Toast.LENGTH_LONG).show()
         }
     }
 
@@ -46,14 +67,14 @@ class RecordActivity : AppCompatActivity() {
         val record = findViewById<TextView>(R.id.textViewRecord)
         database = FirebaseDatabase.getInstance().getReference("PatientRecord")
         database.get().addOnSuccessListener {
-            var sb = StringBuilder()
+            val sb = StringBuilder()
             for(i in it.children){
 
-                var username = i.child("username").value
-                var illness = i.child("illness").value
-                var allergy = i.child("allergy").value
-                var treatment = i.child("treatment").value
-                var id = i.key
+                val username = i.child("username").value
+                val illness = i.child("illness").value
+                val allergy = i.child("allergy").value
+                val treatment = i.child("treatment").value
+                val id = i.key
 
                 sb.append("Record No: $id\nUsername: $username\nIllness: $illness\nAllergy: $allergy\nTreatment: $treatment\n_____________________________\n")
             }
@@ -63,10 +84,23 @@ class RecordActivity : AppCompatActivity() {
         }
     }
 
-    private fun goToEdit() {
+    private fun goToEdit(username: String, name: String, surname: String, number: String) {
         recordNo = findViewById<EditText>(R.id.editTextRecordNo).text.toString()
-        var intent = Intent(this,EditRecordActivity::class.java)
-        intent.putExtra("recordNo",recordNo)
-        startActivity(intent)
+        database = FirebaseDatabase.getInstance().getReference("PatientRecord").child(recordNo)
+        database.get().addOnSuccessListener {
+            if(it.exists()){
+                val intent = Intent(this,EditRecordActivity::class.java)
+                intent.putExtra("recordNo",recordNo)
+                intent.putExtra("username",username)
+                intent.putExtra("name",name)
+                intent.putExtra("surname",surname)
+                intent.putExtra("number",number)
+                startActivity(intent)
+            }else{
+                Toast.makeText(this,"Patient does not exist",Toast.LENGTH_LONG).show()
+            }
+        }.addOnFailureListener {
+            Toast.makeText(this,"failed", Toast.LENGTH_LONG).show()
+        }
     }
 }
